complete the items list and add the edit and delete buttons

+++ lib/features/dash_board/screens/items/presentation/widgets/add_item_container.dart
@@ -64,7 +64,7 @@ class AddItemContainer extends StatelessWidget {
                       if (value == null || value.isEmpty) {
                         return "Price cannot be empty";
                       }
-                      if (int.tryParse(value) == null) {
+                      if (double.tryParse(value) == null) {
                         return "Price must be a number";
                       }
                       return null;
